backend/requirements.txt
airtable-python-wrapper==0.15.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.136.1
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
googleapis-common-protos==1.74.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.13
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.3.2
langchain-openai==1.2.1
langchain-protocol==0.0.12
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.11
langgraph-sdk==0.3.13
langsmith==0.7.36
openai==2.32.0
openpyxl==3.1.5
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.2
proto-plus==1.27.2
protobuf==7.34.1
psycopg2-binary==2.9.12
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pyparsing==3.3.2
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
pyairtable==3.3.0